f32->f64

+++ gateway/src/result_query.rs
@@ -35,7 +35,7 @@ pub struct RankingEntry {
 	/// 本命票数
 	pub first_vote_count: i32,
 	/// 本命率
-	pub first_vote_percentage: f32,
+	pub first_vote_percentage: f64,
 	/// 本命甲醛
 	pub first_vote_count_weighted: i32,
 	/// 所属作品类型
@@ -47,21 +47,21 @@ pub struct RankingEntry {
 	/// 日文名
 	pub name_jpn: String,
 	/// 票数占比
-	pub vote_percentage: f32,
+	pub vote_percentage: f64,
 	/// 本命占比
-	pub first_percentage: f32,
+	pub first_percentage: f64,
 	/// 男性票数
 	pub male_vote_count: i32,
 	/// 男性比例
-	pub male_percentage_per_char: f32,
+	pub male_percentage_per_char: f64,
 	/// 占总体男性比例
-	pub male_percentage_per_total: f32,
+	pub male_percentage_per_total: f64,
 	/// 女性票数
 	pub female_vote_count: i32,
 	/// 女性比例
-	pub female_percentage_per_char: f32,
+	pub female_percentage_per_char: f64,
 	/// 占总体女性比例
-	pub female_percentage_per_total: f32,
+	pub female_percentage_per_total: f64,
 	/// 趋势
 	pub trend: Vec<VotingTrendItem>
 }
